clg chatbot new/requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
python-dotenv==1.0.0
nltk==3.8.1
numpy==1.24.3
scikit-learn==1.3.0
python-dateutil==2.8.2

# Production WSGI servers
gunicorn==21.2.0
waitress==2.1.2

# Optional: Additional production dependencies
# uwsgi==2.0.22
# gevent==23.7.0